Update JDBC.java
Added insertPublisher(Connection), which adds a publisher to the publisher table from user input and returns a string value of the publisher name that was added.
Added replacePublisher(Connection, String newPName) which asks for a publisher name. Replaces all books with that publisher name with the value of newPName.
Updated the menu function to incorporate new functions.